+++ src/rules/index.ts
@@ -12,6 +12,7 @@ import requireComponentIs from './require-component-is';
 import requireVForKey from './require-v-for-key';
 import thisInTemplate from './this-in-template';
 import useVOnExact from './use-v-on-exact';
+import validTemplateRoot from './valid-template-root';
 import validVElse from './valid-v-else';
 import validVElseIf from './valid-v-else-if';
 import validVFor from './valid-v-for';
@@ -31,6 +32,7 @@ export default {
   'require-v-for-key': requireVForKey,
   'this-in-template': thisInTemplate,
   'use-v-on-exact': useVOnExact,
+  'valid-template-root': validTemplateRoot,
   'valid-v-else-if': validVElseIf,
   'valid-v-else': validVElse,
   'valid-v-for': validVFor,