+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   - `this-in-template` ([#6])
     - option `always` currently not supported
   - `use-v-on-exact` ([#20])
+  - `valid-template-root` ([#21])
   - `valid-v-if` ([#28])
   - `valid-v-else-if` ([#24])
   - `valid-v-else` ([#25])
@@ -38,6 +39,7 @@
 [#18]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/18
 [#19]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/19
 [#20]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/20
+[#21]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/21
 [#24]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/24
 [#25]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/25
 [#26]: https://github.com/Shinigami92/eslint-plugin-vue-pug-sfc/issues/26